Version 0.08 Complete! added proper tests for Player Monitor, Attack Manager, Skill Monitor and Status Monitor Fixed Major Bug in Item Manager Fixed Overcharge not being monitored properly (and skills not being used as a consequence) Fixed Channeling not being reacted to properly Fixed Round Manager not Updating Skills Fixed Gems being continuously used once obtained
Bot is now functional. Further Testing will be put into place. Extensive monitoring and data mining of the logs will determine how ready the bot will be.

BUGS:
Items are being stacked when they do not stack in the game, need to have cooldown for certain type of item as well as items themselves.